requirements.txt
Flask==3.0.0
mysql-connector-python==9.0.0
Werkzeug==3.0.1
